requirements.txt
selenium==4.12.0
pytest==7.4.1
pytest-xdist==1.32.0
Faker==1.0.5
webdriver_manager==1.7
allure-pytest==2.6.3
Appium-Python-Client==0.43
pytest-testrail==2.3.3
rootpath==0.1.1
paho-mqtt==1.5.0
paramiko==2.7.1
Jinja2==2.11.2
jsonschema==3.2.0
pyserial==3.4
robotframework-seleniumlibrary==6.1.2
enum34==1.1.10
pywifi==1.1.12
pytz==2021.1
matplotlib
pytest-bdd